+++ src/api-legacy/v1/borehole/stratigraphy/__init__.py
@@ -2,10 +2,7 @@
 
 # Actions
 from bms.v1.borehole.stratigraphy.addbedrock import AddBedrock
-from bms.v1.borehole.stratigraphy.get import GetStratigraphy
-from bms.v1.borehole.stratigraphy.list import ListStratigraphies
 from bms.v1.borehole.stratigraphy.create import CreateStratigraphy
-from bms.v1.borehole.stratigraphy.patch import PatchStartigraphy
 from bms.v1.borehole.stratigraphy.delete import DeleteStratigraphy
 from bms.v1.borehole.stratigraphy.clone import CloneStratigraphy