+++ src/api-legacy/v1/geoapi/__init__.py
@@ -1,5 +1,4 @@
 # -*- coding: utf-8 -*-
 
 # Actions
-from bms.v1.geoapi.wmts import Wmts
 from bms.v1.geoapi.wms import Wms