+++ src/client/src/api-lib/actions/settings.js
@@ -1,4 +1,4 @@
-import { downloadFile, fetch } from "./index";
+import { fetch } from "./index";
 
 export function loadSettings() {
   return fetch("/setting", {
@@ -16,20 +16,4 @@ export function patchSettings(tree, value, key = null) {
     payload.key = key;
   }
   return fetch("/setting", payload);
-}
-
-export function patchEditorSettings(tree, value, key = null) {
-  const payload = {
-    type: "EPATCH",
-    tree: tree,
-    value: value,
-  };
-  if (key !== null) {
-    payload.key = key;
-  }
-  return fetch("/setting", payload);
-}
-
-export function exportDownload(params) {
-  return downloadFile("/setting/export/download", params);
 }