+++ CHANGELOG.md
@@ -9,6 +9,8 @@
 - Upgraded to OpenLayers 8.
 - Upgraded to .NET 8.
 - Upgraded to Node 20.
+- Use numeric input instead of dropdown for top bedrock quality fields.
+- Proxy requests to legacy api through .NET API with authentication.
 
 ## v2.0.452 - 2023-07-05